Replaces build system from Ant to Maven.
Initial work on building the project using Apache Maven,
instead of Apache Ant. Makes easier to upgrade to new
upstream versions of Google BigQuery and AppEngine MapReduce.

Unused code was temporary removed, in favour to support the
builtin Datastore Bakcup to BigQuery tooling, that was also
fixed to search for the proper backup file.